backend/requirements.txt
asgiref==3.11.1
click==8.3.1
colorama==0.4.6
Django==6.0.2
django-stubs==5.2.9
django-stubs-ext==5.2.9
djangorestframework==3.16.1
djangorestframework-stubs==3.16.8
gunicorn==25.1.0
h11==0.16.0
packaging==26.0
sqlparse==0.5.5
types-PyYAML==6.0.12.20250915
typing_extensions==4.15.0
tzdata==2025.3
uvicorn==0.41.0
uvicorn-worker==0.4.0
python-dotenv==1.0.1
google-generativeai==0.5.4
protobuf
psycopg2-binary==2.9.9
